backend/requirements.txt
fastapi
uvicorn
python-multipart
pinecone-client
sentence-transformers
langchain
langchain-community
langchain-huggingface
python-dotenv
requests
